refactor(bench): extract sealed memory source wrapper
